Split sislib.c into its three original translation units
The DOL's float pool repeats 0.0/1.0/640/480/256/-480 and carries four
unreferenced zero bytes at 0x8040C564, neither of which a single TU can
produce. Compiling the three candidate splits shows why: each has its own
.sdata2 pool, and the four bytes are the first TU's .data padded to eight
before the second's.

Split at 0x803A6478 and 0x803A7684. sislib.c keeps the __FILE__ string and
now links: .text, .data, .sdata, .sdata2, .bss and .sbss are byte-identical.
