+++ config/GALE01/splits.txt
@@ -5699,16 +5699,26 @@ sysdolphin/baselib/sobjlib.c:
 	.sdata2     start:0x804DEA70 end:0x804DEAA8
 
 sysdolphin/baselib/sislib.c:
-	.text       start:0x803A5798 end:0x803A949C
-	.data       start:0x8040C490 end:0x8040CD40
+	.text       start:0x803A5798 end:0x803A6478
+	.data       start:0x8040C490 end:0x8040C568
 	.bss        start:0x804D1110 end:0x804D1138
 	.sdata      start:0x804D6390 end:0x804D6398
 	.sbss       start:0x804D7968 end:0x804D7980
-	.sdata2     start:0x804DEAA8 end:0x804DEB10
+	.sdata2     start:0x804DEAA8 end:0x804DEAB8
 
 sysdolphin/baselib/hsd_4D11.c:
 	.bss        start:0x804D1138 end:0x804D3694
 	.sbss       start:0x804D7980 end:0x804D79A0
+
+sysdolphin/baselib/hsd_3A64.c:
+	.text       start:0x803A6478 end:0x803A7684
+	.data       start:0x8040C568 end:0x8040C5A8
+	.sdata2     start:0x804DEAB8 end:0x804DEAC8
+
+sysdolphin/baselib/hsd_3A76.c:
+	.text       start:0x803A7684 end:0x803A949C
+	.data       start:0x8040C5A8 end:0x8040CD40
+	.sdata2     start:0x804DEAC8 end:0x804DEB10
 
 sysdolphin/baselib/sislib_font.c:
 	.data       start:0x8040CD40 end:0x80430B40
