+++ config/GALE01/symbols.txt
@@ -25953,15 +25953,15 @@ lbl_8040C418 = .data:0x8040C418; // type:object size:0x30
 lbl_8040C448 = .data:0x8040C448; // type:object size:0x30
 @244 = .data:0x8040C478; // type:object size:0x16 scope:local data:string
 HSD_SisLib_8040C490 = .data:0x8040C490; // type:object size:0x14 scope:global
+...data.0 = .data:0x8040C490; // type:label scope:local
 HSD_SisLib_8040C4A4 = .data:0x8040C4A4; // type:object size:0x14 scope:global
 HSD_SisLib_8040C4B8 = .data:0x8040C4B8; // type:object size:0x38 scope:global
-HSD_SisLib_8040C4F0 = .data:0x8040C4F0; // type:object size:0x11 scope:global data:string
-HSD_SisLib_8040C504 = .data:0x8040C504; // type:object size:0x9 scope:global data:string
-HSD_SisLib_8040C510 = .data:0x8040C510; // type:object size:0xE scope:global data:string
-HSD_SisLib_8040C520 = .data:0x8040C520; // type:object size:0xF scope:global data:string
-HSD_SisLib_8040C530 = .data:0x8040C530; // type:object size:0x19 scope:global data:string
-HSD_SisLib_8040C54C = .data:0x8040C54C; // type:object size:0x18 scope:global data:string
-HSD_SisLib_8040C564 = .data:0x8040C564; // type:object size:0x4 scope:local
+@68 = .data:0x8040C4F0; // type:object size:0x11 scope:local data:string
+@69 = .data:0x8040C504; // type:object size:0x9 scope:local data:string
+@71 = .data:0x8040C510; // type:object size:0xE scope:local data:string
+@138 = .data:0x8040C520; // type:object size:0xF scope:local data:string
+@273 = .data:0x8040C530; // type:object size:0x19 scope:local data:string
+@274 = .data:0x8040C54C; // type:object size:0x18 scope:local data:string
 jumptable_8040C568 = .data:0x8040C568; // type:object size:0x40 scope:local
 jumptable_8040C5A8 = .data:0x8040C5A8; // type:object size:0x6C scope:local
 jumptable_8040C614 = .data:0x8040C614; // type:object size:0x6C scope:local
@@ -26564,6 +26564,7 @@ particle_list = .bss:0x804D106C; // type:object size:0x44 scope:local align:4 da
 HSD_PSAppSrt_804D10B0 = .bss:0x804D10B0; // type:object size:0x2C scope:global data:byte
 HSD_SObjLib_804D10E0 = .bss:0x804D10E0; // type:object size:0x2C scope:global
 HSD_SisLib_804D1110 = .bss:0x804D1110; // type:object size:0x14 scope:global data:4byte
+...bss.0 = .bss:0x804D1110; // type:label scope:local
 HSD_SisLib_804D1124 = .bss:0x804D1124; // type:object size:0x14 scope:global data:4byte
 hsd_804D1138 = .bss:0x804D1138; // type:object size:0x10 scope:global data:4byte
 hsd_804D1148 = .bss:0x804D1148; // type:object size:0x1200 scope:global data:4byte
@@ -35040,10 +35041,10 @@ HSD_PSDisp_804DEA60 = .sdata2:0x804DEA60; // type:object size:0x8 scope:global a
 @235 = .sdata2:0x804DEA94; // type:object size:0x4 scope:local align:4 data:float
 HSD_SObjLib_804DEA98 = .sdata2:0x804DEA98; // type:object size:0x4 scope:global align:4 data:float
 @249 = .sdata2:0x804DEAA0; // type:object size:0x8 scope:local align:8 data:double
-HSD_SisLib_804DEAA8 = .sdata2:0x804DEAA8; // type:object size:0x4 scope:global align:4 data:float
-HSD_SisLib_804DEAAC = .sdata2:0x804DEAAC; // type:object size:0x4 scope:global align:4 data:float
-HSD_SisLib_804DEAB0 = .sdata2:0x804DEAB0; // type:object size:0x4 scope:global align:4 data:float
-HSD_SisLib_804DEAB4 = .sdata2:0x804DEAB4; // type:object size:0x4 scope:global align:4 data:float
+@139 = .sdata2:0x804DEAA8; // type:object size:0x4 scope:local align:4 data:float
+@140 = .sdata2:0x804DEAAC; // type:object size:0x4 scope:local align:4 data:float
+@264 = .sdata2:0x804DEAB0; // type:object size:0x4 scope:local align:4 data:float
+@265 = .sdata2:0x804DEAB4; // type:object size:0x4 scope:local align:4 data:float
 HSD_SisLib_804DEAB8 = .sdata2:0x804DEAB8; // type:object size:0x4 scope:global align:4 data:float
 HSD_SisLib_804DEABC = .sdata2:0x804DEABC; // type:object size:0x4 scope:global align:4 data:float
 HSD_SisLib_804DEAC0 = .sdata2:0x804DEAC0; // type:object size:0x4 scope:global align:4 data:float
