+++ configure.py
@@ -1916,7 +1916,9 @@ config.libs = [
                 extra_cflags=["-Cpp_exceptions on"],
             ),
             Object(Matching, "sysdolphin/baselib/sobjlib.c"),
-            Object(NonMatching, "sysdolphin/baselib/sislib.c"),
+            Object(Matching, "sysdolphin/baselib/sislib.c"),
+            Object(NonMatching, "sysdolphin/baselib/hsd_3A64.c"),
+            Object(NonMatching, "sysdolphin/baselib/hsd_3A76.c"),
             Object(Matching, "sysdolphin/baselib/sislib_font.c"),
             Object(Matching, "sysdolphin/baselib/hsd_4D11.c"),
             Object(Linkable, "sysdolphin/baselib/hsd_3A94.c"),
